Finished error controls as in issue #40
Fixed a bug with prev. version of controls on Timeline tab. Added an generic error message on the Sidebar.
(If we put messages on separate tabs, we need a separate message for each tab. I do not know why it is so, but it is harder to maintain, so I put the message on the sidebar.)